clothing_detection/requirements.txt
numpy==1.24.3
torch==2.0.1
torchvision==0.15.2
flask
pillow
huggingface-hub
ultralytics==8.0.196
